Remove references to TestCafe
