+++ package.json
@@ -76,7 +76,6 @@
     "@inrupt/solid-client-authn-node": "^1.10.1",
     "@playwright/test": "^1.13.1",
     "@skypack/package-check": "^0.2.2",
-    "@testing-library/testcafe": "^4.2.3",
     "@types/dotenv-flow": "^3.1.0",
     "@types/http-link-header": "^1.0.1",
     "@types/jest": "^27.0.0",
@@ -94,7 +93,6 @@
     "rdf-namespaces": "^1.8.0",
     "rollup": "^2.15.0",
     "rollup-plugin-typescript2": "^0.31.0",
-    "testcafe": "^1.9.4",
     "ts-jest": "^27.0.3",
     "typedoc": "^0.21.0",
     "typedoc-plugin-markdown": "^3.3.0",
